fix(front): do not try to parse directives for every token
The code changes made in `nimconf.nim` focus on optimizing the handling
of directives within the `NimConfParser`. The primary modification
involves the decision to parse directives selectively, rather than
attempting to parse them for every token encountered. This change
addresses inefficiencies in the way directives were previously
processed, which could lead to unnecessary computations when tokens were
not directives.

## API Changes

None.

## Impl Changes

1. The removal of the `confTok` procedure, which was used to handle
   token reading and directive parsing.
2. In its place, a template `nextTok` was introduced to streamline token
   retrieval, cutting down on redundant logic.
3. The parser now checks if a token starts with "@" and only calls
   `parseDirective` when appropriate, thereby reducing overhead,
   improving performance, and fixing a sigsegv.

## Fixes

1. A nimconf statement of form `--ident@[ident]` fails with a SIGSEGV.